Homework 4 spring ioc

+++ resources/db/changelog/sql/07.create-table-payment-card.sql
@@ -3,7 +3,7 @@ create table payment_card
 (
     id            SERIAL PRIMARY KEY,
     cardholder    VARCHAR(128) NOT NULL,
-    number_card   INT UNIQUE   NOT NULL UNIQUE,
+    number_card   BIGINT UNIQUE   NOT NULL UNIQUE,
     valid_thru    VARCHAR(128) NOT NULL,
     security_code INT          NOT NULL
 );
